requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
opencv-python>=4.8.0
streamlit>=1.30.0
grad-cam>=1.5.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
tqdm>=4.65.0
lmdb>=1.4.0  # for nuScenes
imageio>=2.31.0